requirements.txt
streamlit
markdown
pdfkit
python-dotenv
PyMuPDF
PyPDF2
google-generativeai

